Split agent hook files under Swift budget
