+++ Sources/cmuxApp.swift
@@ -7774,10 +7774,7 @@ struct SettingsCardDivider: View {
 struct SettingsCardNote: View {
     let text: String
 
-    init(_ text: String) {
-        self.text = text
-    }
-
+    init(_ text: String) { self.text = text }
     var body: some View {
         Text(text)
             .font(.caption)
